Added some code to read data from our Trac db.
This introduces some models; they were initially generated with inspectdb and
then editied by hand for a while. See the docstring in
django_website/trac/models.py for details.

There's also a db router to make sure that the Trac models get accessed through
the Trac DB correctly.

This is all very rough at the moment, and somewhat constrained by Django's
lack of composite primary key support. But it works for reading data, and
that's enough to start moving forward with it.