A bit more fleshing out of how to handle syncing from github.
The idea is that a manual sync (i.e. the initial one) shouldn't create new
tickets for existing pull requests, but once we handle the webhook we'll
want to create new tickets when new pull requests come in. So there's now
a flag to PullRequest.objects.get_or_create_from_github_dict() telling it
what the right behavior is.