Improved the documentation search form and view.
The search form was designed so that, in the absence of a release ID,
it would yield search results corresponding to the current stable version.

However, there was a bug in the displaying of the form in that
the current release was not selected in the widget.

Another issue is that it was possible to trigger a 500 error by
passing an inexistant release ID to the page.

This commit fixes both those issues and a 404 is now raised when
passing an invalid release ID to the page.