Propagate the fragment identifier in *all* versions (not just dev).
In d48b79618d5d9d50032cc1b585b10e9dd6958044, I added the ability
of the apge to track the fragment identifier in the URL so that
switching between version would preserve it.

However, I overlooked the fact that the block of code I used
was only executed when viewing the dev version of the docs.

This commit makes the fragement identifier tracking available
for all version.