Improved the propagation of URL fragment identifiers between versions.
The commit introduced in da114303d06f413bda604b363c64e07c32b6bee0
only worked if the user didn't navigate to another fragment
after the page had loaded.

With this new approach, the fragment is added to the version
swticher at the last moment which ensures it is up-to-date.