The django community is welcoming and supportive of all people
It's hard enough to avoid thinking and acting in discriminatory
ways when you're aware of your conscious and unconscious tendencies
to do so. It's impossible when you're unaware. Explicitly spelling out a
more comprehensive list of social identities that make us unique helps
us be more cognizant of ways we may accidentally discriminate and make
others feel unwelcome. When others feel unwelcome, the Django
community loses an important perspective. As Daniele Procida so
eloquently reminded us in his keynote address at Djangocon 2014, the
health of our community hinges on our ability to welcome and support
folks who come from a wide variety of backgrounds and situations. He
warned us against the tendency of the software development tribe to be
exlusionary. Let's help prove that it doesn't have to be that way by
publicly stating our goal of being welcoming to all people (and being
explicit about how broadly we define "all people").

FWIW, I showed my list of "discrimination categories" to @evildmp earlier
today and we edited it together. I then decided it would be preferable
to couch the whole statement in the positive. This is why I placed the
statement in the list item about being welcoming rather than placing it
in the list of language and discriminatory behavior in which _not_
to engage.