Abusive behavior elsewhere may affect ability to participate
As discussed with @evildmp and @freakboy3742 on 2014-9-5, if a person
engages in abusive behavior outside of Django community fora, that
may affect their ability to participate within the community (e.g.
they may be placed under special scrutiny or barred altogether). Since
this is already the implicit policy, this simply makes it explicit.

I believe this policy is an important aspect of ensuring that the Django
community is a welcoming and safe place for everyone. If, for example, I
were a woman who knew that someone had made misogynistic statements on
public social networks or in private communication with me, I may not be
comfortable participating in a Django forum, online or in person, in
which that person is present. It would be helpful for me to know that this
code of conduct empowers me to report that person's behavior for the conduct
working group's consideration.