List of identities not comprehensive
This is a reponse to @freakboy3742
(https://github.com/django/djangoproject.com/pull/84/files#r17401314 and
the ensuing discussion). Since it's impossible to construct a
comprehensive list of social identities, I've added the language
"includes, but is not limited to" to make it clear that we want to be
welcoming and supportive of _everyone_.

For some background, I assembled this list based on similar lists I
found in a few places, such as:

- [National Association of Social Workers code of
  ethics](https://www.socialworkers.org/pubs/code/code.asp)
- [University of Michigan Office for Institutional
  Equity](http://www.hr.umich.edu/oie/ndpolicy.html)

I then edited the list with @evildmp.