Added "educational level"
I compared our list with
http://www.ieee.org/about/corporate/governance/p9-26.html and realized
that we were missing something that addressed discrimination
based on educational level. I thought that might be something that comes
up in a technical community such as ours and so is worth including in
this list. IEEE used "matriculation" but I think "educational level,"
while longer, is less ambiguous.