Fix issue #122 - Replace static community stats with actual data.
The data comes from people.djangoproject.com and djangopackages.com and is cached for 12 hours.

This also conditionally enables the page caching middleware during development to be able to use the locmem cache backend.

Thanks to @brutasse for merging https://github.com/brutasse/djangopeople/pull/54 and to @pydanny for being awesome and already providing that data via the Django packages API.