Add update card functionality.
If a card used for a recurring payment expires, Stripe will send a
payment_failed message. We then need some method for the user to update
the subscription with new card details.

This change adds an 'change card details' button to the manage_donations
page which displays a Stripe checkout form and updates the card attached
to the subscription.