Disable Submit button during Ajax processing.
Currently, there is a significant delay between closing the Stripe
checkout form and the page being submitted, while the Ajax code runs.
This is potentially confusing for the user since there is no feedback
that anything is happening. This change adds a visible state change to
the Submit button so that the user is not tempted to click it again.