Cleaned up the updated_docs command
By moving a chunk of logic from the command to a method
on DocumentRelease, it makes the code a bit more readable
and more testable too.
The ElasticSearch indexing was removed from that command
and now needs to be done in a separate step by calling
the update_index command (with the --delete flag).