Added target property and rel="noopener" to external footer link
I was going through the documentation and wanted to check django on twitter via the link on the footer, I did that and when I got to twitter I started doing other things, When I wanted to go back to what I was reading on django I now discovered I was not on a separate Tab. I now decided to fix this.