Block FLoC
Fixes #1091. Use django-permissions-policy to set the `Permissions-Policy` header to `interest-cohort=()`. I thought of disabling more browser features, but figured that was a little risky, so will look at that in the future.

Tested locally with `docker-compose` setup, saw the header set. (It took me a while to figure that the `docker` settings file doesn't import from the `common` settings - why?)