Add a link to Individual membership in the site footer
Finding information about Individual membership can be difficult if one
does not already know where to look. The current path to this page from
the homepage is via the footer at Learn More -> Django Software
Foundation -> Individual Members. This change is in the same spirit as
pull request #1461, which added a link to the Corporate membership page
in the footer for similar discoverability reasons.