Remove Modernizr integration
Modernizr is a feature-detection library that has been a part of this
project for around ten years. Using its CSS classes, this project was
detecting the following browser features:

- Border Radius: https://caniuse.com/border-radius
- Box Shadow: https://caniuse.com/css-boxshadow
- CSS Animations: https://caniuse.com/css-animation
- SVG: https://caniuse.com/svg

Each of these features has been available in modern browsers for many
years, so we no longer need Modernizr.

This patch does the following:

- Removes `modernizr.js`
- Removes PNG logos that were fallbacks for SVG logos
- Refactors SCSS that was pivoting on Modernizr CSS classes (`mdzr-*`)