Use CSS instead of JavaScript for floating warning
Some of the functionality of the floating warning on unsupported and
in-development documentation versions was implemented using JavaScript.
This is longer necessary with `position: sticky` in CSS. This feature is
widely-supported: https://caniuse.com/css-sticky.

Other than setting `position: sticky` and removing the related
JavaScript, this patch does the following:

- Simplify the CSS so that the class on all floating warning elements
  implements the common rules and the CSS for the IDs on the elements
  implements the colors
  - A `z-index` value was added to the floating warning element because
    the banner element in the header has one specified
- Move the floating warning element to the top of the page in the HTML
  - Remove the `body_extra` block that was only being used by the
    floating warning element
  - Add a new `before_header` block for the floating warning element