Removed bower
`bower` is maintained but not recommended for use by its developers. Our
`bower.json` file now only includes two dependencies: `jquery` and
`jquery-flot`. At this point, `bower` is not doing much to serve the
project.

This is a redo of 98e20a08b69597d9e0c081c20809007b3e363407 which
was reverted because it led to errors in production when running
collectstatic.

Refs #1827