Run `prettier` on JavaScript via `pre-commit`
This patch removes the global `djangoproject/static/*` exclude rule in
`.pre-commit-config.yaml` to enable `prettier` to run on this project's
JavaScript files. The `djangoproject/static/js/lib/` directory is now
excluded in the `prettier` hook because there is no value in
reformatting a minified JavaScript file.

The following changes were also made:

- Add a `.prettierrc` file to the root of the project to control
  JavaScript formatting
- Run `prettier` on all existing JavaScript that should be formatted