Restored serving of spinx _images
This is a partial revert of efc51e848d870d4c7675f6636b9893324ba3bc3f.
Due to an unfortunate conflict in the nginx configuration of
the server after the deploying of the commit mentioned above,
the _images directory ended up still being served by Django
which added a / to the end of the image path, using a permanent
HTTP redirection (because of settings.APPEND_SLASH).
The solution (hopefully temporary) is to restore the serving
of images by Django while we figure out how to configure
nginx to keep serving the _images and handled the old cached
redirections with a trainling slash.