Update the fundraising goal amount from 200000 to 300000
This is per the fundraising working group’s discussions. The amount is pretty arbitrary, we just know we want more funding this year, so +100k felt like a reasonable place to aim.

The only place this figure is used is on the [fundraising page](https://www.djangoproject.com/fundraising/), to display the absolute goal amount and % towards completion.