Simplify handling of release roadmap image class (#2046)
Due to the fact that a cookie can override the user's system setting for
dark mode, the class on the release roadmap image was being managed by
JavaScript.

However, we can detect all cases in CSS since the `setTheme` JavaScript
function has updated the `data-theme` attribute on the `body` tag in all
cases.

This patch also groups all CSS related to the release roadmap image
together in `_dark-mode.scss`.