WIP support of including blog in the website search.
The blog results should have a property of whether it is included in
the search results. We should also limit the blogs that are searchable
for a version of Django based on the support end. This will allow us
to limit the inclusion of blog posts in the search based on the time
the entry was created, keeping the search results relevant to that
version of Django.