Moved and simplified `fundraising-heart.js`
jQuery integration has been removed. It was only used for a single
element selection call and a single data attribute access call.

The file has been removed from the module system, put in the general
JavaScript directory, and added to the one page that uses it.

The actual `Heart` code has not been changed.