Isolated `dashboard` app JavaScript
This patch moves the `dashboard` app's JavaScript inclusion from the
project's `require.js` structure to the app itself.

This was achieved with a new block in the base template called
`javascript`. The rest of the site gets the default JavaScript includes,
and the `dashboard` app overrides the block to include the JavaScript
specific to it. The `dashboard` app benefits by not knowing about
JavaScript related to the rest of the site and vice versa.

None of the actual `dashboard` JavaScript code was changed. It was only
adapted to the new structure.

- Added `javascript` block to base template
- Added `dashboard_javascript` block to `dashboard` base template
- Included `dashboard` JavaScript directly in `dashboard` templates
- Removed `dashboard`-related entries in `main.js`
- Remove references to `jquery.flot.min.js` from base template
  - This library is only used by the `dashboard` app
- Removed `dashboard-*` CSS classes only used by `require.js`
- Removed `require.js` structure in `dashboard` JavaScript files
- Simplified structure of `utils.js`