Add top `CorporateMember`s to `docs` sidebar
This was requested by the fundraising working group.

I abstracted the top `CorporateMember` display from the `releases` app
sidebar into a new inclusion tag called `top_corporate_members` and
included it in its original place as well as in the `docs` sidebar as
requested.

Uploaded images for `CorporateMember`s may not display in local
development environments do to the way media files paths work with
subdomains.